% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{style_nb}
\alias{style_nb}
\title{Style the notebook using the \code{styler} package.}
\usage{
style_nb(notebook_file, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{notebook_file}{the notebook file name}

\item{overwrite}{overwrite existing ipynb file (default: FALSE)}

\item{verbose}{print information about styling (default: TRUE)}
}
\description{
Style the notebook using the \code{styler} package.
}
